requirements.txt
Django==5.2.8
django-notifications-hq==1.8.3
djangorestframework==3.16.1
python-telegram-bot==22.5
